Move `GcAllocator` to library/alloc
This means that programs can import `std::gc::GcAllocator` without
needing to use the `#[feature(rustc_private)]` attribute.
