+++ library/alloc/src/gc.rs
@@ -45,6 +45,7 @@ use crate::boxed::Box;
 use std::boxed::Box;
 
 use core::{
+    alloc::{AllocError, Allocator, GlobalAlloc, Layout},
     any::Any,
     borrow,
     cmp::Ordering,
@@ -62,18 +63,76 @@ use core::sync::atomic::{self, AtomicU64};
 #[cfg(not(no_global_oom_handling))]
 use core::gc::ReferenceFree;
 
-use boehm::GcAllocator;
-
 #[cfg(test)]
 mod tests;
-
-#[unstable(feature = "gc", issue = "none")]
-static ALLOCATOR: GcAllocator = GcAllocator;
 
 #[cfg(profile_gc)]
 static FINALIZERS_REGISTERED: AtomicU64 = AtomicU64::new(0);
 #[cfg(profile_gc)]
 static FINALIZERS_COMPLETED: AtomicU64 = AtomicU64::new(0);
+
+#[derive(Debug)]
+pub struct GcAllocator;
+
+unsafe impl GlobalAlloc for GcAllocator {
+    #[inline]
+    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
+        unsafe { boehm::GC_malloc(layout.size()) as *mut u8 }
+    }
+
+    #[inline]
+    unsafe fn dealloc(&self, ptr: *mut u8, _: Layout) {
+        unsafe {
+            boehm::GC_free(ptr);
+        }
+    }
+
+    #[inline]
+    unsafe fn realloc(&self, ptr: *mut u8, _: Layout, new_size: usize) -> *mut u8 {
+        unsafe { boehm::GC_realloc(ptr, new_size) as *mut u8 }
+    }
+}
+
+unsafe impl Allocator for GcAllocator {
+    #[inline]
+    fn allocate(&self, layout: Layout) -> Result<NonNull<[u8]>, AllocError> {
+        unsafe {
+            let ptr = boehm::GC_malloc(layout.size()) as *mut u8;
+            let ptr = NonNull::new_unchecked(ptr);
+            Ok(NonNull::slice_from_raw_parts(ptr, layout.size()))
+        }
+    }
+
+    unsafe fn deallocate(&self, _: NonNull<u8>, _: Layout) {}
+}
+
+impl GcAllocator {
+    pub fn force_gc() {
+        unsafe { boehm::GC_gcollect() }
+    }
+}
+
+pub fn init() {
+    unsafe { boehm::GC_init() }
+}
+
+/// Returns true if thread was successfully registered.
+pub unsafe fn register_thread(stack_base: *mut u8) -> bool {
+    unsafe { boehm::GC_register_my_thread(stack_base) == 0 }
+}
+
+/// Returns true if thread was successfully unregistered.
+pub unsafe fn unregister_thread() -> bool {
+    unsafe { boehm::GC_unregister_my_thread() == 0 }
+}
+
+pub fn suppress_warnings() {
+    unsafe { boehm::GC_set_warn_proc(&boehm::GC_ignore_warn_proc as *const _ as *mut u8) };
+}
+
+pub fn thread_registered() -> bool {
+    unsafe { boehm::GC_thread_is_registered() != 0 }
+}
 
 struct GcBox<T: ?Sized>(T);
 
@@ -279,7 +338,7 @@ impl<T> Gc<T> {
         }
 
         unsafe {
-            ALLOCATOR.register_finalizer(
+            boehm::GC_register_finalizer_no_order(
                 self.ptr.as_ptr() as *mut u8,
                 Some(finalizer::<T>),
                 null_mut(),
@@ -292,7 +351,15 @@ impl<T> Gc<T> {
     #[unstable(feature = "gc", issue = "none")]
     pub fn unregister_finalizer(&mut self) {
         let ptr = self.ptr.as_ptr() as *mut GcBox<T> as *mut u8;
-        ALLOCATOR.unregister_finalizer(ptr);
+        unsafe {
+            boehm::GC_register_finalizer(
+                ptr,
+                None,
+                ::core::ptr::null_mut(),
+                ::core::ptr::null_mut(),
+                ::core::ptr::null_mut(),
+            );
+        }
     }
 }
